#9695: fix upgrade
